final commit - changed the entity id to not autogen and to Integer istead of Long, fixed the dao to int and added the findbyid to restservice class

+++ src/java/dao/TeacherDAO.java
@@ -23,14 +23,14 @@ public class TeacherDAO implements TeacherDAOInterface {
     }
 
     @Override
-    public void removeTeacher(Long id) {
+    public void removeTeacher(int id) {
 
         em.remove(em.find(Teacher.class, id));
 
     }
 
     @Override
-    public Teacher findById(Long id) {
+    public Teacher findById(int id) {
         return em.find(Teacher.class, id);
     }
